feat: implement redis

+++ src/modules/users/users.service.ts
@@ -1,5 +1,6 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException
 } from '@nestjs/common'
@@ -14,12 +15,12 @@ import {
   IFindOptions,
   IPaginationOptions
 } from './interfaces/options.interface'
-import { UserRepository } from './repositories/user.repository'
+import { UserRedisRepository } from './repositories/redis/user.redis.repository'
 
 @Injectable()
 export class UsersService {
   constructor(
-    private readonly userRepository: UserRepository,
+    private readonly userRepository: UserRedisRepository,
     private readonly addressesService: AddressesService,
     private readonly envService: EnvService
   ) {}
@@ -54,14 +55,24 @@ export class UsersService {
   async updateUserById(
     tx: PrismaClientTransaction,
     id: string,
+    sub: string,
     { address, password, ...dto }: UpdateUserDto
   ): Promise<UserEntity> {
+    const user = await this.findUserById(tx, id, {
+      deletedAt: false,
+      disabledAt: false
+    })
+
+    if (!user || sub !== user.id) {
+      throw new ForbiddenException('No permission to edit.')
+    }
+
     if (dto.email) {
       await this.verifyUserExistsByEmail(
         tx,
         dto.email,
         { deletedAt: undefined, disabledAt: undefined },
-        id
+        sub
       )
     }
 
@@ -70,48 +81,93 @@ export class UsersService {
         tx,
         dto.login,
         { deletedAt: undefined, disabledAt: undefined },
-        id
+        sub
       )
     }
 
     return await this.userRepository.updateUserById(tx, id, dto)
   }
 
-  async countUsers(tx: PrismaClientTransaction, options: IPaginationOptions) {
-    return await this.userRepository.countUsers(tx, options)
-  }
-
   async findAllUsersForPagination(
     tx: PrismaClientTransaction,
     options: IPaginationOptions
-  ) {
+  ): Promise<UserEntity[]> {
     return await this.userRepository.findAllUsersForPagination(tx, options)
   }
 
-  async findAllUsers(tx: PrismaClientTransaction, options: IFindOptions) {
-    return await this.userRepository.findAllUsers(tx, options)
-  }
+  async disableUserById(
+    tx: PrismaClientTransaction,
+    id: string,
+    sub: string
+  ): Promise<UserEntity> {
+    const user = await this.findUserByIdOrThrow(tx, id, {
+      deletedAt: false,
+      disabledAt: false
+    })
 
-  async disableUserById(tx: PrismaClientTransaction, id: string) {
+    if (!user || sub !== user.id) {
+      throw new ForbiddenException('No permission to edit.')
+    }
+
     return await this.userRepository.disableUserById(tx, id)
   }
 
-  async enableUserById(tx: PrismaClientTransaction, id: string) {
+  async enableUserById(
+    tx: PrismaClientTransaction,
+    id: string,
+    sub: string
+  ): Promise<UserEntity> {
+    const user = await this.findUserByIdOrThrow(tx, id, {
+      deletedAt: false,
+      disabledAt: undefined
+    })
+
+    if (!user || sub !== user.id) {
+      throw new ForbiddenException('No permission to edit.')
+    }
+
     return await this.userRepository.enableUserById(tx, id)
   }
 
-  async deleteUserById(tx: PrismaClientTransaction, id: string) {
+  async deleteUserById(
+    tx: PrismaClientTransaction,
+    id: string,
+    sub: string
+  ): Promise<UserEntity> {
+    const user = await this.findUserByIdOrThrow(tx, id, {
+      deletedAt: false,
+      disabledAt: false
+    })
+
+    if (!user || sub !== user.id) {
+      throw new ForbiddenException('No permission to edit.')
+    }
+
     return await this.userRepository.deleteUserById(tx, id)
   }
 
   // VERIFY //
+
+  async findAllUsers(
+    tx: PrismaClientTransaction,
+    options: IFindOptions
+  ): Promise<UserEntity[]> {
+    return await this.userRepository.findAllUsers(tx, options)
+  }
+
+  async countUsers(
+    tx: PrismaClientTransaction,
+    options: IPaginationOptions
+  ): Promise<number> {
+    return await this.userRepository.countUsers(tx, options)
+  }
 
   async verifyUserExistsByEmail(
     tx: PrismaClientTransaction,
     email: string,
     options: IFindOptions,
     currentUserId?: string
-  ) {
+  ): Promise<void> {
     const user = await this.findUserByEmail(tx, email, options)
 
     if (user && user.id !== currentUserId) {
@@ -124,7 +180,7 @@ export class UsersService {
     login: string,
     options: IFindOptions,
     currentUserId?: string
-  ) {
+  ): Promise<void> {
     const user = await this.findUserByLogin(tx, login, options)
 
     if (user && user.id !== currentUserId) {
@@ -136,7 +192,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     id: string,
     options: IFindOptions
-  ) {
+  ): Promise<void> {
     if (await this.findUserById(tx, id, options)) {
       throw new ConflictException('ID already registered.')
     }
@@ -146,7 +202,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     email: string,
     options: IFindOptions
-  ) {
+  ): Promise<UserEntity> {
     const user = await this.findUserByEmail(tx, email, options)
     if (!user) throw new NotFoundException('User not found.')
     return user
@@ -156,7 +212,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     login: string,
     options: IFindOptions
-  ) {
+  ): Promise<UserEntity> {
     const user = await this.findUserByLogin(tx, login, options)
     if (!user) throw new NotFoundException('User not found.')
     return user
@@ -166,7 +222,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     id: string,
     options: IFindOptions
-  ) {
+  ): Promise<UserEntity> {
     const user = await this.findUserById(tx, id, options)
     if (!user) throw new NotFoundException('User not found.')
     return user
@@ -176,7 +232,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     email: string,
     options: IFindOptions
-  ) {
+  ): Promise<UserEntity> {
     return await this.userRepository.findUserByEmail(tx, email, options)
   }
 
@@ -184,7 +240,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     login: string,
     options: IFindOptions
-  ) {
+  ): Promise<UserEntity> {
     return await this.userRepository.findUserByLogin(tx, login, options)
   }
 
@@ -192,7 +248,7 @@ export class UsersService {
     tx: PrismaClientTransaction,
     id: string,
     options: IFindOptions
-  ) {
+  ): Promise<UserEntity> {
     return await this.userRepository.findUserById(tx, id, options)
   }
 }
